gp: avoid failure if /dev/shm not present
The `Sys.is_directory fpath` call fails with
```
Fatal error: exception Sys_error("fpath: No such file or directory")
```
when the file is not present. The suggested solution
prevents this issue.

A better solution would also check if the folders are writable,
and if not create a different local temp dir on the fly, but
I think it is overkill in this case.

Signed-off-by: Marcello Seri <marcello.seri@gmail.com>